feat: Allow xmls in subdirs by using os.walk()
We switch to os.walk() for finding XML files, to allow the usage of
subdirectories, for example to sort the plugin registry by category,
such as cheat plugins, story plugins, or deprecated plugins.

Or in the case of extremely large plugin repositories, to split all
XMLs into a set of folders depending on their initial letter, as
GitHub truncates all file directory listings to 1,000 files.