feat(cli): Remind player to update their registries
Every time the registry is updated, the time that the update occured is
set in the SQLite database of naevpm.

When the CLI is run, this timestamp
is checked and if it is more than 1
week in the past, the CLI will remind the player to update their plugin
registries, even giving the command to do so.